Improve index.html
I'm mostly doing this to get an idea for what I may eventually have to do for the other pages, if I don't re-write their HTML entirely.
In the end, this is valid HTML, at least both for validators and browsers.  The semantics could be way better.  (other than the BR elements and META element not being self-closing, and for being  labeled HTML 4.01, it's also valid XHTML)
It's still ugly.  Good TODOs would be to eliminate nbsp, and to use actual list elements for the table of contents.